# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video archive API backed by Supabase
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Recording {
  id: string;
  /** @nullable */
  channel_id?: string | null;
  username: string;
  filename: string;
  timestamp: string;
  /** @nullable */
  room_title?: string | null;
  tags?: string[];
  /** @nullable */
  viewers?: number | null;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  framerate?: number | null;
  /** @nullable */
  filesize?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  sprite_url?: string | null;
  /** @nullable */
  embed_url?: string | null;
  /** @nullable */
  preview_url?: string | null;
  /** @nullable */
  instance_id?: string | null;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface RecordingListResponse {
  data: Recording[];
  total: number;
  page: number;
  limit: number;
}

export interface Performer {
  username: string;
  recording_count: number;
  /** @nullable */
  latest_thumbnail?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  latest_timestamp?: string | null;
}

export interface PerformerProfile {
  username: string;
  recording_count?: number;
  /** @nullable */
  gender?: string | null;
  recordings: Recording[];
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface SiteStats {
  total_recordings: number;
  total_performers: number;
  total_tags: number;
  total_size_bytes: number;
  /** @nullable */
  newest_recording?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface ReactionCount {
  likes: number;
  dislikes: number;
  /** @nullable */
  user_reaction?: string | null;
}

export interface ReactionInput {
  recording_id: string;
  type: string;
  session_id: string;
}

export interface Comment {
  id: number;
  recording_id: string;
  /** @nullable */
  parent_id?: number | null;
  author: string;
  content: string;
  deleted?: boolean;
  likes: number;
  user_liked?: boolean;
  created_at: string;
  replies?: Comment[];
}

export interface CreateCommentInput {
  recording_id: string;
  author: string;
  content: string;
  session_id: string;
}

export interface CreateReplyInput {
  author: string;
  content: string;
  session_id: string;
}

export interface SessionInput {
  session_id: string;
}

export interface CommentLikeResult {
  likes: number;
  liked: boolean;
}

export type ListRecordingsParams = {
page?: number;
limit?: number;
search?: string;
/**
 * Comma-separated list of tags to filter by
 */
tags?: string;
gender?: string;
username?: string;
resolution?: string;
sort?: ListRecordingsSort;
};

export type ListRecordingsSort = typeof ListRecordingsSort[keyof typeof ListRecordingsSort];


export const ListRecordingsSort = {
  newest: 'newest',
  oldest: 'oldest',
  largest: 'largest',
  popular: 'popular',
} as const;

export type ListRelatedRecordingsParams = {
id: string;
limit?: number;
};

export type GetReactionsParams = {
recording_id: string;
session_id?: string;
};

export type ListCommentsParams = {
recording_id: string;
sort?: ListCommentsSort;
session_id?: string;
};

export type ListCommentsSort = typeof ListCommentsSort[keyof typeof ListCommentsSort];


export const ListCommentsSort = {
  top: 'top',
  new: 'new',
  old: 'old',
} as const;