# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video archive API backed by Supabase
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  recording_id: string;
  /** @nullable */
  parent_id?: number | null;
  author: string;
  content: string;
  deleted?: boolean;
  likes: number;
  user_liked?: boolean;
  created_at: string;
  replies?: Comment[];
}