# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video archive API backed by Supabase
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  session_id: string;
}